Brings the ML-compiler branch to main:
- faithful realize-seed CPU breakthrough (beautiful_mnist beats tinygrad)
- divergent-consumer partial-realize (transformer attention forward)
- conv backward fixes (SHRINK swizzle, EXPAND memory, multi-axis reduce)
- grad correctness: maxpool tie-split, DETACH stop-grad, shared-extent reduce
- optimizers: on-graph Adam/AdamW bias correction, LARS, Muon (Newton-Schulz)

Verified at merge time: 0 conflicts, 0 deletions, change set disjoint from the
in-flight ATP working tree. thvm side built clean (test_grad 68/68, byte-clean
faithful parity); WL/ATP paclet builds + TFindProof runs on the merged tree.

+++ wl/THVMLink/CSource/thvmlink.c
@@ -1900,7 +1900,7 @@ EXTERN_C DLLEXPORT int thvm_wl_kernel_apply_opt(WolframLibraryData l, mint a,
     MArgument_setInteger(res, 0);
     return LIBRARY_NO_ERROR;
   }
-  KOpt opt = { (u8)op, (u8)axis, arg };
+  KOpt opt = { (u8)op, axis, arg };
   int ok = tile_anno_apply_opt(&KERNELS[kid], opt);
   if (ok) {
   }
